AuthController fixed.
Paths matching rn.
